Refactor UTs by moving them into their own folders.
